tests: Extend MediaWikiIntegrationTestCase as these write to the DB
Extending PHPUnit's own TestCase is strongly discouraged, not least
because it means that database writes will actually write to the DB
on the developer's machine, rather than a test database for the run
and so potentially damaging data. As of recently, MediaWiki will do
nothing in this case when detected, but instead throw an error; the
patch here should get the repo back to working state again.

Depends-On: Ic5e3d6138a094e6e6bbeb7cad1354263d76664f7
Change-Id: I980cfd796c6e71f6044d6ba5bd718ec3c93b5fe0